#000: Shubham : Remove the long to double widening deprecation warnings.

+++ csw-event/csw-event-client/src/multi-jvm/scala/event.client/perf/commons/PerfSubscriber.scala
@@ -133,7 +133,7 @@ class PerfSubscriber(
       payloadSize,
       histogram,
       eventsReceived,
-      totalTime,
+      totalTime.toDouble,
       outOfOrderCount,
       avgLatency()
     )
